Add files via upload
Task 1 
A Rotary Club is organizing fund raiser selling take-away lunch sets. Price for an adult lunch set is $20 and price for kids set is $12. Customers can buy more than one adult set and kid set. Side orders can be added on at a separate price which includes $4 for salad and $5 for dessert.

Task 2
Write a program that asks for the names of three runners and the time, in minutes, it took each of them to finish a race. The program should display the names of the runners in the order that they finished.

Task 3
You have been asked to develop an application using a switch statement that allows theatre employee to enter a day of the week and then display the cost of a theatre ticket for the day. Incorporate switch and if..else statement to apply the following conditions in the application.
Day Age Cost

Monday - Thursday Under 10 Free 10 – 65 RM10 Over 65 Free

Friday - Sunday Under 10 RM10 10 – 65 RM20 Over 65 RM10